+++ src/main/java/act/inject/SessionValueLoader.java
@@ -20,6 +20,7 @@ package act.inject;
  * #L%
  */
 
+import act.apidoc.ApiManager;
 import act.app.ActionContext;
 import org.osgl.$;
 import org.osgl.inject.BeanSpec;
@@ -46,6 +47,9 @@ public class SessionValueLoader implements ValueLoader<Object> {
 
     @Override
     public Object get() {
+        if (ApiManager.inProgress()) {
+            return null;
+        }
         ActionContext ctx = ActionContext.current();
         E.illegalStateIf(null == ctx, "Not in request handling context");
         return $.convert(ctx.session(sessionKey)).to(targetType);